requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
psycopg2-binary
openai
python-dotenv
opencv-python
pytesseract
Pillow
reportlab  # (for PDF export later)
